Multiple Adjustments To Animations, Muzzle Light, Enemy UV's
Damage tweaks for Kyle: Piercing/Ricochet and Explosion damage Should be fixed.
Adjusted spawn timing to better match incoming animations.
Adjusted death timing so it is variable and can spawn a gut bag at different times depending on a delay.
Put a light on the muzzle for when you are shooting that changes colour to match your bullets.
Fixed enemies never going into their idle animation
Added in Jonah's turning variable on the spider walks, they are very creepy now
Tiled the enemies UV's in code based on their bullet element
Started work on our collision fixes, not quite there but it is a good building block for tomorrow